fastapi-gen/pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "fastapi-gen"
version = "0.1.0"
description = "CLI tool for generating FastAPI projects with clean architecture"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "jinja2>=3.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
fastapi-gen = "fastapi_gen.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
